requirements.txt
# Core scientific stack
pandas==2.2.*
numpy==1.26.*
scikit-learn==1.5.*
lightgbm==4.5.*
optuna==3.6.*
scipy==1.13.*

# Persistence / IO
pyarrow==16.0.*
joblib==1.4.*
PyYAML==6.0.1

# Streamlit UI
streamlit==1.35.*
plotly==5.22.*

# Data collection (optional)
python-binance==1.0.*
pycoingecko==3.1.*
requests>=2.32.0

# Misc
loguru==0.7.*